tests/requirements.txt
redis
pytest
dnspython
